python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
rake-nltk==1.0.6
anthropic==0.5.0
openai==1.3.5
pytest==7.4.3